Dev2/feature/create add Room and set relative

+++ 7.3.0/aspnet-core/src/EMS.Core/Authorization/Schedules/Schedule.cs
@@ -18,7 +18,7 @@ namespace EMS.Authorization.Schedules
         public long WorkShiftId { get; set; }
         public WorkShift WorkShift { get; set; }
         [ForeignKey("Room")]
-        public long RoomId { get; set; }
+        public int RoomId { get; set; }
         public Room Room { get; set; }
     }
 }
